Revert "Update stats_create.js"
This reverts commit d79b7f19819074b85f3434299c98ff1afa1c1b3d.

+++ workers/mysql/stats_create.js
@@ -23,13 +23,13 @@ function find_power(data) {
   var power = 0;
 
   if (found1) {
-    power = parseFloat(found1[0]);
+    power = parseInt(found1[0]);
   } else if (found2) {
-    power = parseFloat(found2[0]);
+    power = parseInt(found2[0]);
   } else if (found3) {
-    power = parseFloat(found3[0]) * 1000;
+    power = parseInt(found3[0]) * 1000;
   } else if (found4) {
-    power = parseFloat(found4[0]) * 1000 * 1000;
+    power = parseInt(found4[0]) * 1000 * 1000;
   }
 
   return power.toFixed(2);
